move element-type dispatch for connection candidates in a new module
